backend/requirements.txt
fastapi==0.128.7
uvicorn==0.40.0
SQLAlchemy==2.0.46
python-dotenv==1.2.1
requests==2.32.5
